Allow the user to specify the order in which dns entries are created.
Having run into some issues with the use of both a prefixed hostname and
the base libvirt domain name in DNS, I'd like to be able to specify the
order in which the hostnames in an entry are created. To explain, if the
entry is ordered:

10.122.0.2 foo test-foo test-foo.example.com

then when the host does a reverse lookup it will decide that it's
hostname should be foo, where I'd like it to be test-foo instead. I'd
still like to be able to have the foo entry in there so that I can ssh
to the bare domain name, but I'd like the name returned to be the full
hostname (test-foo).

To resolve this I've implemented support for specifying the order in
which the hostnames are added to the dns entry. The default behaviour is
to use the bare domain name only, but if the prefix and/or domain name
are supplied the order can be specified however you want.