Make the hostname search robust against leftover files.
If there's an old .leases file (which can happen if you update a host in place
from Trusty to Xenial, as it turns out), it'll fail to find the new MAC and
fall over in a very un-informative way. This fixes/works around that problem by
trying the status file code path if nothing can be found in the leases file.