Improve correctness of the Ec2 API implementation.
There's been an issue with cloud-init failing to successfully run its
puppet module, and I think I tracked it down to a problem where the
instance-id data was being treated as a /list/ rather than as a single
value. Looking at the metadata response on an openstack cloud it looks
like there's no trailing new line on single values, so I went in and
fixed /that/, but at the same time I've gone in and changed the
implementation of a number of things to better match the way that the
actual Ec2 meta-data interface is described, particularly around public
keys.