Allow the user to supply extra data to templates.
This change adds a new section to the config file called
'template-data', which contains key=value pairs that will be made
available to the template engine (and hence to the userdata templates).
The data is made available using the unmodified key, so that (for
example):

[template-data]
foo=bar

would make {{foo}} available for use in the template, with value "bar".

Values are always passed as strings, with no conversion of any kind - if
any type conversion is required it needs to be done in python code in
the template.