Make EC2 version support dynamic.
This removes the singular 'md_base' config option, and replaces it with
a 'service.ec2_versions' option. This option is a comma-separated list
of URL bases that represent an EC2 version, i.e:

[service]
ec2_versions = 2009-04-04, latest

results in the following versions response:

2009-04-04/
latest/

with all the associated paths functioning as expected. There is no way
to provide different features to the different paths, but this provides
some additional flexibility.