Remove remaining code from original repository.
Also update a bunch of other stuff including documentation.

This moves to using the entry_points feature of setuptools, thus
removing the need for a directly specified bin/mdserver. It also removes
the old libvirt hooks and the md-test.sh script that haven't been
relevant for a long time. Finally, it updates the README file and moves
it over to a markdown format for better prettiness.