Fix issues handling default templates from a file.
For some reason this seems to cause the Bottle.template() function to
produce a unicode object rather than a plain str object - there's no
issue with the hard-coded default template string, which is why I didn't
spot this error.

The "fix" I've implemented here should work without any issues on both
python 2 and 3 - it's a null operation on python 3, but on python 2 it
converts the unicode to a str.

If you really need to use a unicode default template, run this with
python 3 and it should Just Work(tm).