Add support for a persistent database.
The database is updated by sending a POST request with the contents of
the domain XML definition of the instance to the `/instance-upload` path
on the mdserver. The name, uuid and mds network MAC address are recorded
in the database, along with timestamps for the first time it was seen
and the most recent time it was seen.