Add proper support for configuring dnsmasq.
This is quite an extensive change, which introduces support for creating
DHCP and DNS host files based on the contents of the current database,
as well as creating a dnsmasq configuration file that makes use of these
files.

In addition it properly handles IP address allocation on the network,
using a very simple iterative approach to find the lowest unused IP
address, updating the database and then from there updating the dnsmasq
host files.

Finally, the example mdserver.conf file has been updated to provide
documentation of all the new configuration that's needed to support this
new functionality.