Update and complete install setup.
This is a significant change from the old setup, where the setup.py
script was responsible for installing lots of system integration stuff
(init scripts, systemd unit files, config files, etc) as well as the
core application. Since I've discovered that this is really not
something that's well supported by setuptools and the like (and really
shouldn't be supported) I've moved away from relying on setuptools
towards a two-part installation: setup.py to do the application install,
and then a separate script to do the systems integration (or at least,
to provide a quick start type of setup of the systems integration
components.

This is all done by removing the `data_files` section from the setup.py
script, and adding a `system-integration.sh` script which puts the
supplied files in what should be the right places, as well as creating
a user for dnsmasq to do its thing as.

This change also cleans up a number of other obsolete things, including
removing the /etc/default file - if we're expecting the user to be
responsible for their sytem integration files, they can edit things if
they want to use non-default values for anything of any interest.