Make IP address allocation random rather than sequential.
This is to minimse the risk of having a migrating instance coming over
from another host with an address that clashes with one of the existing
instances - this would be a very real risk if the addresses were
allocated linearly, since most would be allocated at the bottom of the
address range. Spreading them out across the whole range minimises that
as far as possible (using a /16 makes it even less risky).

This also includes a number of other code cleanups and some additional
tests, along with a proper tox based test runner.