Add support for multiple configuration files.
This adds support for loading additional configuration files, which can
extend or override the contents of the main configuration file (as
specified on the command line). Extra config files are specified in the
`[_include_]` section of the main file, and are loaded after the main
file. Additional files can override anything in the main file, with
each additional file overriding values from previous files - the last
value specified wins.

Two varieties of additional config can be specified: directories,
specified as a comma-separated list in the value of the `directories`
key, from which all the *.conf files will be loaded; and a list of
individual files, likewise specified as a comma-separated list in the
value of the `files` key. Entries in each list are loaded in the order
they appear in the list, with all directories being processed before any
files are processed.

Also included in this change is an update to the `tests/run-test.sh`
script and associated infrastructure - this makes use of the new config
file support.