Add support for metadata embedded in domain XML.
This feature extracts metadata from the domain XML, using the mdserver
namespace, and makes it available in the database. This is currently
only used to support specifying a userdata template prefix which will
override the default search patterns; it may be used for other things
later on.

This change also includes a string of small clean ups and tweaks, which
were identified during development, as well as tweaks to the test suite.
However, this includes significant improvements to the test suite, and
associated clean ups of the code being tested.