Fix crash in the database metadata handling code.
This was due to not making sure the format of the database entries was
as expected, leading to using the wrong check for the existence of
metadata associated with an entry. The fix was to set things up properly
in the first place.

In addition, I added code to make sure this is more robust against
similar things in the future, including re-using an old database file
with new code. I also added tests to exercise all this code.