Add support for exporting running config.
This will allow other tools to know how to coordinate with the mdserver
process - initially I'm thinking a database query tool, but probably
other things will be able to use it.

The request is limited to the local host only, for obvious reasons, and
filters out a number of potentially sensitive config items for safety's
sake. It may need to move to an allow list rather than a deny list,
though . . .