Properly test the interactions between mdserver and dnsmasq.
This introduces a fake dnsmasq process that writes its pid to the
dnsmasq pidfile (as configured by mdserver), sets up a signal handler
for SIGHUP, and on receiving the signal it checks the existence and
state of the generated files. This allows us to test almost all of this
part of the code - all that's missing is actually running dnsmasq for
real (rather than just in its `--test` syntax checking mode).