Big reworking of database handling, and various other bits.
In aid of supporting additional database backends, I've moved the
database implementing into its own submodule, and broken the old
implementation up into a core Database class, and a JsonDatabase child
class that provides the JSON backend implementation. That should make it
reasonably easy to replace the JSON backend with something else (I've
been eyeing etcd, but I've recently started considering the possibility
of redis).

I've also changed the database schema significantly - it's no longer
just a simple list of entries, it includes a metadata section that
records information about the mdserver instances that are managing it.
This is to help support multiple mdservers using the same database,
which is something I'm planning to use in my work setup. I have yet to
implement locking to go with this - part of that is because I'd like to
use portalocker to support the locking, and that supports redis as a
locking backend, and if I'm considering using redis as an alternative
backend then it'd make sense to do that at the same time as I added the
additional portalocker requirement . . .

Finally, I tossed the old setup.cfg, moving everything into
pyproject.toml. This fixed a number of issues with the packaging that
I'd been hitting repeatedly, and it means I can have a more meaningful
version string. I still need to figure out a way to handle my desire to
support the release date as well - for now I'm just doing that manually
. . .