Disable interpolation when loading config files.
Bottle's ConfigDict uses the ExtendedInterpolation class when creating
its ConfigParser objects, which means that any $ and { characters need
to be escaped. Given we expect people to specify at least a passwd hash,
which uses $ characters as part of the standard format, this is a right
pain.

There's not a whole lot of use for interpolation in these config files,
so the simplest thing by far is to just disable it entirely.